Reformats and removes some dead code
